additions to handle #1

+++ setup.py
@@ -20,21 +20,21 @@ def msg(text, verbose = False):
         print(text)
 
 # execute subprocess
-def exec(command, verbose = False):
+def exec(command, verbose = False, shell = False):
     if verbose:
-        subprocess.run(shlex.split(command))
+        subprocess.run(shlex.split(command), shell=shell)
     else:
-        subprocess.run(shlex.split(command), stdout=subprocess.DEVNULL)
+        subprocess.run(shlex.split(command), stdout=subprocess.DEVNULL, shell=shell)
 
 # execute subprocess in another working dir
-def exec_cwd(command, wd, verbose = False):
+def exec_cwd(command, wd, verbose = False, shell = False):
     if verbose:
-        subprocess.run(shlex.split(command), cwd=(wd))
+        subprocess.run(shlex.split(command), cwd=(wd), shell=shell)
     else:
-        subprocess.run(shlex.split(command), cwd=(wd), stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
+        subprocess.run(shlex.split(command), cwd=(wd), stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, shell=shell)
 
 # grabs all required git repos
-def clone_repos(verbose = False):
+def clone_repos(verbose = False, shell = False):
     print('\nCloning all required repositories:')
     base = 'git clone https://github.com/'
     repos = [ \
@@ -49,13 +49,13 @@ def clone_repos(verbose = False):
         exec(base+repo, verbose)
 
 # installs all dependancies via pip
-def install_pytorch(verbose = False):
+def install_pytorch(verbose = False, shell = False):
     print('\nInstalling Pytorch (this may take a few minutes):')
     cmd = 'conda install -y pandas pytorch torchvision torchaudio cudatoolkit=11.3 -c pytorch'
     exec(cmd, verbose)
 
 # installs all dependancies via pip
-def install_dependencies(verbose = False):
+def install_dependencies(verbose = False, shell = False):
     print('\nInstalling all required dependencies:')
     base = 'pip install --no-input '
     packages = [ \
@@ -85,7 +85,7 @@ def install_dependencies(verbose = False):
         exec(base+package, verbose)
 
 # perform real-esrgan setup
-def setup_esrgan(verbose = False):
+def setup_esrgan(verbose = False, shell = False):
     print('\nSetting up Real-ESRGAN:')
     cmd = 'curl -L -o Real-ESRGAN/experiments/pretrained_models/RealESRGAN_x4plus.pth -C - https://github.com/xinntao/Real-ESRGAN/releases/download/v0.1.0/RealESRGAN_x4plus.pth'
     msg('   downloading pre-trained models...', verbose)
@@ -96,22 +96,35 @@ def setup_esrgan(verbose = False):
     msg('   running ESRGAN setup script...', verbose)
     exec_cwd(cmd, cwd + os.path.sep + 'Real-ESRGAN', verbose)
 
+# for debugging
+def test(verbose = False, shell = False):
+    print('\nTesting dummy call:')
+    base = 'thiswillfail'
+    exec(base, verbose)
+
 # create config.txt file
 def create_config():
-    print('\nCreating config.txt containing default settings...')
-    shutil.copy2('config-default.txt', 'config.txt')
+    if not os.path.exists('config.txt'):
+        print('\nCreating config.txt containing default settings...')
+        shutil.copy2('config-default.txt', 'config.txt')
+    else:
+        print('\nExisting config.txt found, leaving it intact...')
 
 # updates these repos to the latest version
-def update():
+def update(shell = False):
     print('\nChecking for updates:')
     cmd = 'git pull'
     cwd = os.getcwd()
 
-    print('updating dream-factory...')
-    exec(cmd, True)
-
-    print('updating stable-diffusion...')
-    exec_cwd(cmd, cwd + os.path.sep + 'stable-diffusion', True)
+    try:
+        print('updating dream-factory...')
+        exec(cmd, True, shell)
+        print('updating stable-diffusion...')
+        exec_cwd(cmd, cwd + os.path.sep + 'stable-diffusion', True, shell)
+    except FileNotFoundError:
+        print('\nUnable to find git executable - try re-running update with the --shell option, e.g.:')
+        print('   python setup.py --shell --update\n')
+        exit(-1)
 
 # entry point
 if __name__ == '__main__':
@@ -130,6 +143,12 @@ if __name__ == '__main__':
     )
 
     parser.add_argument(
+        "--shell",
+        action='store_true',
+        help="enable shell calls; try this if setup cannot find your executables"
+    )
+
+    parser.add_argument(
         "--update",
         action='store_true',
         help="updates a previous installation"
@@ -137,23 +156,29 @@ if __name__ == '__main__':
 
     opt = parser.parse_args()
     verbose = opt.verbose
+    shell = opt.shell
 
     if opt.update:
         if os.path.exists('stable-diffusion'):
             # user is requesting an update
-            update()
+            update(shell)
         else:
             print('\nNo previous installation detected; aborting update.')
             print('Run without the --update switch first.\n')
     else:
         if opt.force or not os.path.exists('stable-diffusion'):
             # either no previous installation or the user is forcing re-install
-            clone_repos(verbose)
-            install_pytorch(verbose)
-            install_dependencies(verbose)
-            setup_esrgan(verbose)
-            create_config()
+            try:
+                clone_repos(verbose, shell)
+                install_pytorch(verbose, shell)
+                install_dependencies(verbose, shell)
+                setup_esrgan(verbose, shell)
+            except FileNotFoundError:
+                print('\nUnable to find executable - try re-running setup with the --shell option, e.g.:')
+                print('   python setup.py --shell --force\n')
+                exit(-1)
 
+            create_config()
             checkpoint_path = 'stable-diffusion\models\ldm\stable-diffusion-v1'
             if not os.path.exists(checkpoint_path):
                 os.makedirs(checkpoint_path)
